Refactored types into types/ dir
