+++ firestore/angular-rewrite/src/app/submit-review-modal/submit-review-modal.component.ts
@@ -18,7 +18,7 @@ import { Component, ViewEncapsulation, inject, Inject } from "@angular/core";
 import { Firestore, collection, addDoc } from "@angular/fire/firestore";
 import { MatDialogRef, MAT_DIALOG_DATA } from "@angular/material/dialog";
 
-import { Rating } from "../review-list/ratings";
+import { Rating } from "../../../types/ratings";
 
 @Component({
   selector: "app-submit-review-modal",
@@ -37,7 +37,7 @@ export class SubmitReviewModalComponent {
   constructor(
     public dialogRef: MatDialogRef<SubmitReviewModalComponent>,
     @Inject(MAT_DIALOG_DATA) private data: string
-  ) {}
+  ) { }
 
   onCancelClick() {
     this.dialogRef.close();
