+++ firestore/angular-rewrite/types/ratings.ts
