+++ firestore/angular-rewrite/types/restaurant.ts
